refactor: Rename files for CodingZone
