+++ frontend/src/index.js
@@ -4,12 +4,12 @@ import { BrowserRouter, Route, Routes, useLocation } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Main from './pages/HomePage';
 import ArticleMain from './pages/Article/ArticleMain';
-import CodingMain from './pages/Coding-zone/CodingMain';
+import CodingMain from './pages/Coding-zone/CodingZoneMain';
 import EditPage from './pages/Article/EditPage';
-import CodingClassRegist from './pages/Coding-zone/CodingClassRegist';
-import Codingzone_Manager from './pages/Coding-zone/Codingzone_Manager';
-import Codingzone_Attendence from './pages/Coding-zone/Codingzone_Attendence';
-import Codingzone_All_Attend from './pages/Coding-zone/Codingzone_All_Attend';
+import CodingClassRegist from './pages/Coding-zone/CodingZoneRegist';
+import CodingZoneAttendanceAssistant from './pages/Coding-zone/CodingZoneAttendanceAssistant';
+import CodingZoneMyAttendance from './pages/Coding-zone/CodingZoneMyAttendance';
+import CodingZoneAttendanceManager from './pages/Coding-zone/CodingZoneAttendanceManager';
 import NavBar from './components/NavBar';
 import Footer from './components/Footer';
 import CreatePage from './pages/Article/CreatePage';
@@ -32,9 +32,9 @@ const App = () => {
         <Route path="/article-main/:articleNum" element={<ShowPage />} />
         <Route path="/coding-zone" element={<CodingMain />} />
         <Route path="/coding-zone/coding-class-regist" element={<CodingClassRegist />} />
-        <Route path="/coding-zone/Codingzone_Manager" element={<Codingzone_Manager />} />
-        <Route path="/coding-zone/Codingzone_Attendance" element={<Codingzone_Attendence />} />
-        <Route path="/coding-zone/Codingzone_All_Attend" element={<Codingzone_All_Attend />} />
+        <Route path="/coding-zone/Codingzone_Manager" element={<CodingZoneAttendanceAssistant />} />
+        <Route path="/coding-zone/Codingzone_Attendance" element={< CodingZoneMyAttendance/>} />
+        <Route path="/coding-zone/Codingzone_All_Attend" element={<CodingZoneAttendanceManager/>} />
       </Routes>
       {location.pathname !== '/' && <Footer />}
     </>
